+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Added `default_fts_config` dynamic configuration parameter for changing text search settings.
 - Added `is_group` field to Mastodon API `Status` entity.
 - Added API method for creating groups.
+- Added `filter` parameter to `/api/v1/groups/followed` API endpoint.
 
 ### Changed